Update AVCodec to work with Homebrew ffmeg
Fix the header imports. I don't know what distributions use the ffmpeg folder,
but the fairly recent version provided by Homebrew uses the libavcodec and
libavformat directories.

Update the source to use the *2 variants of the audio/video decode functions.
The non-suffixed variants are completely gone in the Homebrew version of
ffmpeg and the *2 variants themselves are still deprecated.

Squelch a warning about fprintf() being used without a format string.